refactor: address PR review feedback for CN v0.75 bump
- Move BACKUP_DATABASE_DUMP_NOT_FOUND error code (SOLO-4080) to after
  SOLO-4079 so codes appear sequentially in file order
- Remove dead write in updateBlockStreamPropertiesForMode: the BOTH
  branch set streamWrappedRecordBlocks=true but every call site
  immediately overwrites it to false via ensureWrappedRecordBlocksDisabled
- Consolidate get_latest_mirror_block_number, wait_for_mirror_block_progress,
  and wait_for_mirror_block_count_progress into helper.sh, removing
  duplicates from launch_network.sh and solo_smoke_test.sh
- Extract RSA address-book JSON building into Helpers.buildRsaAddressBookJson
  so tasks.ts and block-node.ts share one implementation
- Update refreshBlockNodeRsaBootstrapState task title to surface pod
  restart side-effect that was invisible to operators
