+++ .github/workflows/script/launch_network.sh
@@ -158,65 +158,6 @@ show_service_ips() {
   kubectl get svc -n "${namespace}" network-node1-svc network-node2-svc -o custom-columns=NAME:.metadata.name,CLUSTER-IP:.spec.clusterIP,CREATED:.metadata.creationTimestamp
 }
 
-get_latest_mirror_block_number() {
-  local mirror_url="${1:-http://127.0.0.1:38081}"
-  local response=""
-
-  response=$(curl -sfS \
-    -H 'Cache-Control: no-cache, no-store, must-revalidate' \
-    -H 'Pragma: no-cache' \
-    -H 'Expires: 0' \
-    "${mirror_url}/api/v1/blocks?limit=1&order=desc" || true)
-
-  node -e '
-const fs = require("fs");
-const input = fs.readFileSync(0, "utf8");
-try {
-  const data = JSON.parse(input);
-  const value = Number(data.blocks?.[0]?.number);
-  console.log(Number.isFinite(value) ? value : -1);
-} catch {
-  console.log(-1);
-}
-' <<< "${response}"
-}
-
-wait_for_mirror_block_progress() {
-  local label="${1}"
-  local previous_block="${2:--1}"
-  local max_attempts="${3:-90}"
-  local sleep_seconds="${4:-2}"
-  local latest_block=-1
-  local minimum_block=$((previous_block + 1))
-
-  echo "$(date '+%Y-%m-%d %H:%M:%S') - Waiting for mirror block ingestion (${label}), minimum block ${minimum_block}" >&2
-  for ((attempt = 1; attempt <= max_attempts; attempt++)); do
-    latest_block=$(get_latest_mirror_block_number)
-    if [[ "${latest_block}" -ge "${minimum_block}" ]]; then
-      echo "$(date '+%Y-%m-%d %H:%M:%S') - Mirror block ingestion ready (${label}): latest block ${latest_block}" >&2
-      echo "${latest_block}"
-      return 0
-    fi
-
-    echo "Mirror block ingestion not ready (${label}) [attempt=${attempt}/${max_attempts}, latest=${latest_block}, minimum=${minimum_block}]" >&2
-    sleep "${sleep_seconds}"
-  done
-
-  echo "Timed out waiting for mirror block ingestion (${label}); latest=${latest_block}, minimum=${minimum_block}" >&2
-  return 1
-}
-
-wait_for_mirror_block_count_progress() {
-  local label="${1}"
-  local previous_block="${2:--1}"
-  local required_new_blocks="${3:-1}"
-  local max_attempts="${4:-90}"
-  local sleep_seconds="${5:-2}"
-  local minimum_previous_block=$((previous_block + required_new_blocks - 1))
-
-  wait_for_mirror_block_progress "${label}" "${minimum_previous_block}" "${max_attempts}" "${sleep_seconds}"
-}
-
 # Restart relay after upgrade and refresh port-forwards.
 refresh_relay_network_config() {
   local namespace="${1}"
@@ -736,7 +677,7 @@ solo one-shot falcon deploy \
 SKIP_IMPORTER_CHECK=true
 .github/workflows/script/solo_smoke_test.sh "${SKIP_IMPORTER_CHECK}"
 
-wait_for_mirror_block_progress "source deployment after one-shot" -1 90 2 > /dev/null
+wait_for_mirror_block_progress "source deployment after one-shot" -1 1 90 2 > /dev/null
 source_block_after_one_shot="$(get_latest_mirror_block_number)"
 echo "$(date '+%Y-%m-%d %H:%M:%S') - Source mirror block before consensus upgrade: ${source_block_after_one_shot}"
