+++ src/core/errors/error-code-registry.ts
@@ -213,11 +213,11 @@ export const ErrorCodeRegistry: Record<string, string> = {
   BACKUP_INPUT_PATH_NOT_FOUND: 'SOLO-4073',
   BACKUP_INPUT_MUST_BE_ZIP: 'SOLO-4074',
   BACKUP_NO_LOG_FILES: 'SOLO-4075',
-  BACKUP_DATABASE_DUMP_NOT_FOUND: 'SOLO-4080',
   FLAG_INPUT_FAILED: 'SOLO-4076',
   CONFIRMATION_REQUIRED: 'SOLO-4077',
   VALUES_FILE_NOT_FOUND: 'SOLO-4078',
   VALUES_FILE_PARSE_FAILED: 'SOLO-4079',
+  BACKUP_DATABASE_DUMP_NOT_FOUND: 'SOLO-4080',
 
   // 5xxx - System / Environment: kubectl, DNS, permissions, timeouts
   RESOURCE_NOT_FOUND: 'SOLO-5001',
